requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
distro==1.9.0
executing==2.2.0
faker==37.1.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
openai==1.78.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.4
pydantic-core==2.33.2
pyfiglet==1.0.2
pygments==2.19.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
sniffio==1.3.1
sqlalchemy==2.0.40
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.23.0
